チャット機能と要約機能をまとめる
