+++ base.py
@@ -10,15 +10,24 @@ from langchain.schema import (
 #AIMessage  # ChatGPTの返答
 from langchain.callbacks import get_openai_callback
 
+import requests
+from bs4 import BeautifulSoup
+from urllib.parse import urlparse
 
-def init_page():
+def init_select_mode():
     st.set_page_config(
-        page_title="ChatGPTを流用したアプリ",
+        page_title="ChatGPTを流用したwebサイト",
         page_icon="🤗"
     )
-    st.header("ChatGPTを流用したアプリ 🤗")
+    mode = st.sidebar.radio("使う機能を選択:", ("gpt_chat", "web_summary"))
+    if mode == "gpt_chat":
+        mode_name = "gpt_chat"
+        st.header("ChatGPT機能 🤗")
+    else:
+        mode_name = "web_summary"
+        st.header("webサイトの要約 🤗")
     st.sidebar.title("Options")
-
+    return mode_name
 
 def init_messages():
     clear_button = st.sidebar.button("Clear Conversation", key="clear")
@@ -27,7 +36,6 @@ def init_messages():
             SystemMessage(content="システムメッセージ：あなたの手助けを行います")
         ]
         st.session_state.costs = []
-
 
 def select_model():
     model = st.sidebar.radio("Choose a model:", ("GPT-3.5", "GPT-4"))
@@ -42,37 +50,98 @@ def select_model():
 
     return ChatOpenAI(temperature=temperature, model_name=model_name)
 
+def get_url_input():
+    url = st.text_input("URL: ", key="input")
+    return url
+
+def validate_url(url):
+    try:
+        result = urlparse(url)
+        return all([result.scheme, result.netloc])
+    except ValueError:
+        return False
+
+def get_content(url):
+    try:
+        with st.spinner("Fetching Content ..."):
+            response = requests.get(url)
+            soup = BeautifulSoup(response.text, 'html.parser')
+            # fetch text from main (change the below code to filter page)
+            if soup.main:
+                return soup.main.get_text()
+            elif soup.article:
+                return soup.article.get_text()
+            else:
+                return soup.body.get_text()
+    except:
+        st.write('something wrong')
+        return None
+
+def build_prompt(content, n_chars=300):
+    return f"""以下はとある。Webページのコンテンツである。内容を{n_chars}程度でわかりやすく要約してください。
+            ========
+            {content[:1000]}
+            ========
+            日本語で書いてね！
+            """
 
 def get_answer(llm, messages):
     with get_openai_callback() as cb:
         answer = llm(messages)
     return answer.content, cb.total_cost
 
-
 def main():
-    init_page()
+    mode_name = init_select_mode()
 
     llm = select_model()
     init_messages()
 
-    # ユーザーの入力を監視
-    if user_input := st.chat_input("聞きたいことを入力してね！"):
-        st.session_state.messages.append(HumanMessage(content=user_input))
-        with st.spinner("ChatGPT is typing ..."):
-            answer, cost = get_answer(llm, st.session_state.messages)
-        st.session_state.messages.append(AIMessage(content=answer))
-        st.session_state.costs.append(cost)
+    if mode_name == "gpt_chat":
+        if user_input := st.chat_input("聞きたいことを入力してね！"):
+            st.session_state.messages.append(HumanMessage(content=user_input))
+            with st.spinner("ChatGPT is typing ..."):
+                answer, cost = get_answer(llm, st.session_state.messages)
+            st.session_state.messages.append(AIMessage(content=answer))
+            st.session_state.costs.append(cost)
+        messages = st.session_state.get('messages', [])
+        for message in messages:
+            if isinstance(message, AIMessage):
+                with st.chat_message('assistant'):
+                    st.markdown(message.content)
+            elif isinstance(message, HumanMessage):
+                with st.chat_message('user'):
+                    st.markdown(message.content)
+            else:  # isinstance(message, SystemMessage):
+                st.write(f"System message: {message.content}")
+    else:
+        container = st.container()
+        response_container = st.container()
 
-    messages = st.session_state.get('messages', [])
-    for message in messages:
-        if isinstance(message, AIMessage):
-            with st.chat_message('assistant'):
-                st.markdown(message.content)
-        elif isinstance(message, HumanMessage):
-            with st.chat_message('user'):
-                st.markdown(message.content)
-        else:  # isinstance(message, SystemMessage):
-            st.write(f"System message: {message.content}")
+        with container:
+            url = get_url_input()
+            is_valid_url = validate_url(url)
+            if not is_valid_url:
+                st.write('Please input valid url')
+                answer = None
+            else:
+                # URLのパース
+                content = get_content(url)
+                if content:
+                    # パースした内容と質問文書を作成
+                    prompt = build_prompt(content)
+                    st.session_state.messages.append(HumanMessage(content=prompt))
+                    with st.spinner("ChatGPT is typing ..."):
+                        answer, cost = get_answer(llm, st.session_state.messages)
+                    st.session_state.costs.append(cost)
+                else:
+                    answer = None
+        if answer:
+            with response_container:
+                st.markdown("## Summary")
+                st.write(answer)
+                st.markdown("---")
+                st.markdown("## Original Text")
+                st.write(content)
 
     costs = st.session_state.get('costs', [])
     st.sidebar.markdown("## Costs")
